Option arguments, private member refactoring, parser changes, usability
Adds `.accepts()` to Option, and `.accepts()` to Command that will pass through
to the last added option. Added `.required()` to Command, which makes the last
added Command or Option required.

Changed the constructor of Option to allow specifying a type of value to
accept. If this is omitted, the option acts as a boolean flag. This is a
breaking change.

The parser now recognizes "--" as a signal to put all remaining arguments in
the string into the "_" property of the results.

Updated the README to be a little more comprehensive, and to reflect the
changes in this commit. Tests have been expanded.